Importation of PIL is now more likely to succeed
Depending of the way PIL has been installed, it is possible that the
importation of the whole PIL module fails, but the importation of its
submodules succeed. If the importation with PIL.submodule doesn't work,
then pyqrtcode now try to import its submodule directly.

This bug appears when PIL is installed with easy_install and not by the
Windows installer nor the package from various distribution directories.

For more information, see
http://stackoverflow.com/questions/8863917/importerror-no-module-named-pil